Add special __DEFAULT__ gemset to include all default gems
"Global" gemset is not used when no gemset is active.
A rbenv user usually installs gems into the default directory, not "global" gemset.
This option make it easy to reuse these gems in the default directory.

Current implementation get gempath by

    RBENV_GEMSETS=" " gem env gempath

but it is ugly.  Please modify if there is a better way