Fixed gemset list if $RBENV_ROOT was not set.
There was an issue where $RBENV_ROOT is not set, and the list script did not work. The rbenv project added their own function to find the root which was added in place of just the $RBENV_ROOT env var. This passes off the function of finding this root to rbenv and allows rbenv-gemset list to work in both instances now.