Added the support for placing the special `-global` and `-standard` options into the .rbenv-gemset file.
These gemset ```-global` disables the usage of the global gemset and the gemset ``-standard`` disables the usage of the standard gem installation directory. Code works with space, newline or space-newline mix separated gemset list. @jf decided that we should not use an internal command, so the code is duplicated in ``libexec/rbenv-gemset-active`` and ``etc/rbenv.d/exec/gemset.bash``. I guess this is ok, as long as we do not need option parsing in other parts of rbenv-gemset. The changes enable us to use rbenv-gemset to sandbox gem installations.

Commit is a proposal to resolve jf/rbenv-gemset#59